Add `_vi_module.py`
Move `VELOVAE` and `DecoderVELOVI` to separate module file.
